refactor: 재사용가능한 함수를 다른 파일로 이동시키고 import하여 사용
