+++ navCategories.js
@@ -1,4 +1,4 @@
-import { createTagTemplate, createLiListTemplate } from "./util/createTag.js";
+import { createTagTemplate, createLiListTemplate } from "./util/util.js";
 import { categories } from "./data/contents.js";
 
 const element = document.querySelector(".main__navigation__menu");
